requirements.txt
aiogram==3.1.1
aiosqlite==0.19.0
python-dotenv==1.0.0
geopy==2.3.0